Refactor BitSequence to encode stuff bit differently to format them
The BitSequence internally kept a list of booleans to represents the
bits of the sequence. Now they are represented with integers (defined
in _SeqBit), where 0 and 1 are the regular bit values and 2 and 3
are the stuffed-bit values (2 == stuff-bit-0, 3 == stuff-bit-1).

This is then used in BitSequence.toBinString*() functions to format the
stuff bits differently than the other bits, using the new optional
parameters of string prefix and string suffix of the stuff bit
string representation, the simplest being "<b>" before and "</b>"
after the stuff bit.